added controllers, authMiddleware, responseToken
