+++ server/model/db_handler.js
@@ -1,8 +1,9 @@
-import Sequelize from "sequelize";
+// import Sequelize from "sequelize";
+const Sequelize = require("sequelize");
 
 const sequelize = new Sequelize("IMS", "root", "My$ql@wb", {
   host: "localhost",
   dialect: "mysql",
 });
 
-export default sequelize;
+module.exports = sequelize;
